Put Relay.Marathon.State in its own file too
